backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-allauth==65.14.3
django-cors-headers==4.3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Faker==24.8.0
fastapi==0.115.12
pydantic==2.11.3
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
sqlparse==0.5.5
jsonschema==4.26.0
jsonschema[format]==4.26.0
pytest-cov==7.0.0
joblib==1.4.2
numpy==1.26.4
scikit-learn==1.5.0
Django==6.0.2
python-dotenv==1.0.1
sqlalchemy==2.0.30
python-dotenv>=1.0.0
pandas>=2.0.0
